feature: start connecting proposals and proposal details
