+++ app/src/pages/guild/proposal_details_page.tsx
@@ -1,13 +1,48 @@
-import React, { useState } from 'react'
+import moment from 'moment'
+import React, { useEffect, useState } from 'react'
+import { RouteComponentProps, useHistory } from 'react-router'
 
 import { ProposalDetailsView } from '../../components/guild/proposal_details_view_container'
+import { useConnectedWeb3Context } from '../../contexts'
+import { useGraphMarketMakerData, useGuildProposals } from '../../hooks'
+import { Proposal } from '../../services/guild'
 
-export const ProposalDetailsPage = () => {
+interface RouteParams {
+  id: string
+}
+
+export const ProposalDetailsPage = (props: RouteComponentProps<RouteParams>) => {
+  const { networkId } = useConnectedWeb3Context()
+
+  const history = useHistory()
+
+  const { proposals } = useGuildProposals()
+
+  const proposalId = props.match.params.id
   const [isScalar, setIsScalar] = useState(false)
+  const [proposal, setProposal] = useState<Proposal>()
+
+  // eslint-disable-next-line
+  const { marketMakerData } = useGraphMarketMakerData(proposal ? proposal.description : '', networkId)
+
+  useEffect(() => {
+    if (proposals.length) {
+      const proposal = proposals.find(proposal => proposal.id === proposalId)
+      setProposal(proposal)
+    }
+  }, [proposals, proposalId])
+
+  const proposalEndDate = proposal && new Date(proposal.endTime.toNumber() * 1000)
+  // eslint-disable-next-line
+  const formattedProposalEndDate = moment(proposalEndDate).fromNow(true)
+
+  const back = () => history.push('/guild')
+
   //logic
   const dummyDataPassed = {
     amount: '500.00 OMN',
     apy: '360%',
+    back,
     duration: '32 days',
     marketDetails:
       'What will the June 2021 CME/Globex S&P500 e-mini terminate at? https://www.cmegroup.com/trading/equity-index/us-index/e-mini-sandp500_quotes_globex.html',
